+++ index.js
@@ -3,6 +3,8 @@ const fs = require('fs')
 const path = require('path')
 
 const debug = require('debug')('procmonrest')
+const isPlainObject = require('is-plain-obj')
+const mergeObjects = require('lodash.merge')
 const terminate = require('tree-kill')
 
 /**
@@ -101,9 +103,26 @@ class Procmonrest {
 
     const privateData = {
       cmd: options.command || 'npm start',
+      env: {
+        NODE_ENV: 'test'
+      },
       log: false,
       pattern: options.waitFor,
       ref: options.reference
+    }
+
+    if (options.envars) {
+      if (!isPlainObject(options.envars)) {
+        throw new Error('If specified, the "envars" option must be a plain object containing only string values.')
+      }
+
+      Object.keys(options.envars).forEach((key) => {
+        if (typeof options.envars[key] !== 'string') {
+          throw new Error('If specified, the "envars" option must be a plain object containing only string values.')
+        }
+
+        privateData.env[key] = options.envars[key]
+      })
     }
 
     if (options.saveLogTo === undefined) {
@@ -192,6 +211,7 @@ class Procmonrest {
       privateData.cmd,
       {
         cwd: workingDirectory,
+        env: mergeObjects({}, process.env, privateData.env),
         shell: true,
         stdio: 'pipe'
       }
@@ -209,6 +229,8 @@ class Procmonrest {
           .filter(line => line.length > 0)
 
         lines.forEach((line) => {
+          debug('STDOUT: %s', line)
+
           if (privateData.log) {
             privateData.log.stream.write(`STDOUT: ${line}\n`)
           }
@@ -230,7 +252,7 @@ class Procmonrest {
             }
 
             privateData.ready = true
-            resolve()
+            resolve(line)
           }
         })
       })
@@ -241,7 +263,10 @@ class Procmonrest {
             .toString()
             .split(/\r?\n/)
             .filter(line => line.length > 0)
-            .forEach(line => privateData.log.stream.write(`STDERR: ${line}\n`))
+            .forEach((line) => {
+              debug('STDERR: %s', line)
+              privateData.log.stream.write(`STDERR: ${line}\n`)
+            })
         }
       })