+++ CHANGELOG.md
@@ -4,11 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+* The `start` method now resolves to the entire line that matched `waitFor`.
+* The ability to specify environment variables for the child process.
+
+
 ## [1.1.0] - 2021-06-13
 
 ### Added
 
-* The `saveLogTo` option can still be used to specify the location of the log file; however, it will now default to the same location as the file that calls `new Procmonrest()`.
+* The `saveLogTo` option can still be used to specify the location of the log file; however, it now has a default value that is the same as the location of the file that calls `new Procmonrest()`.
 
 
 ## [1.0.0] - 2021-04-06